Cowserver dev (#53)
* Rename config item to man_use_cowmaster

- handle non Linux trying to use cowmaster, in setup.py
- add timeout waiting for cowmaster on launch

* Handle difference between game server and cowmaster in client connections

Since they are different clients, we know check if cowmaster or game server before running functions in packet reception

* Resolve incorrect overwrite of cmdline argument dictionary item

Should be an insert, not a direct overwrite. Was causing a misconfiguration in game servers

* cowserver > cowmaster (typo)

* reinstate creation of game server objects.

Cow master is seperate to game servers now, so there is no crossover between them previously affecting cowserver fitting into max game server pool count.

* resolve missing logic to handle game server id vs cowmaster ID assignment

* FIXME: Remove calls to 'check_for_restart_required' when using cowmaster

Reason:
 - when cowmaster is used, servers are forked instead of manager creating processes
 - it seems each servers cmdline arguments are identical to one another when forked
 - since we previously checked for inconsistencies in the commandline arguments to detect changes, this causes an infinite loop of shutting down / starting up servers when forking

IMPACT:
 - when changing configuration values, servers wont automatically restart and apply them.

* drafting function to get a process by incoming client connection over TCP

* fix typo in attribute name

* missing "self" in new function definition

* Resolve killing of duplicate cowmaster servers.

* Update logging to INFO for killing cowmaster duplicate

* Update references to game_server to include support for cowmaster in packet parser

Some parsing affects both game server and cowmaster, some just affects game server.

* Additional conditionals in parser for game_server vs cowmaster

* Additional conditionals to packet parser game_server vs cowmaster

* Send LOGGER to cowmaster packet parser

* Add unsetting of _pid values when a server TCP connection is lost

* remove condition in set_client_connection

Now it will always get the latest PID.
This is probably a cleaner fix to supersede the previous commit.

* Add CowMaster process hook to handle unexpected closure

Also removed early return in game_server.start_server()
Now all initialization should be completed with a forked or non forked server

* Unused variable

* missing await expression

* add proc lookup to cowmaster when it connects

also add unset_client_connection to remove pid reference

* add missing attribute into init

* re-implement server restart on config item change

Modify which cmdline arguments are compared.

* Add check before terminating cowmaster

+++ cogs/game/cow_master.py
@@ -17,11 +17,12 @@ class CowMaster:
         self.port = port
         self.id = 0
         self.global_config = global_config
+        self.enabled = False
 
         self.client_connection = None
         self.cowmaster_cmdline = get_cowmaster_configuration(self.global_config.get("hon_data"))
 
-        self.game_manager_parser = GameManagerParser(self.id)
+        self.game_manager_parser = GameManagerParser(self.id, logger = LOGGER)
         
         # state variables
         self._started = None
@@ -30,8 +31,10 @@ class CowMaster:
         self.status_received = asyncio.Event()
 
         self.game_state = GameState()
-        self.reset_game_state()
+        self.reset_cowmaster_state()
         self.game_state.add_listener(self.on_game_state_change)
+
+        asyncio.create_task(self.monitor_process())
     
     async def fork_new_server(self, game_server):
         if not self.client_connection:
@@ -40,24 +43,35 @@ class CowMaster:
         await self.client_connection.send_packet(game_server.get_fork_bytes(), send_len=True)
 
     async def start_cow_master(self):
+        """
+            Linux only feature, the cow master is used to preload resources for each available map type.
+            The cow master can then be commanded to "fork" new game servers, off existing resources and RAM.
+            This results in instant server startup times and some significantly less RAM usage overall 
+        """
         cmdline_args = MISC.build_commandline_args(self.cowmaster_cmdline, self.global_config, cowmaster = True)
-        if MISC.get_os_platform() == "win32":
-            os.environ["APPDATA"] = str(self.global_config['hon_data']['hon_artefacts_directory'])
-            os.environ["USERPROFILE"] = str(self.global_config['hon_data']['hon_home_directory'])
-            DETACHED_PROCESS = 0x00000008
-            exe = subprocess.Popen(cmdline_args,close_fds=True, creationflags=DETACHED_PROCESS)
-        else:
-            exe = subprocess.Popen(cmdline_args,close_fds=True,start_new_session=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+        exe = subprocess.Popen(cmdline_args,close_fds=True,start_new_session=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
         self._pid = exe.pid
         self._proc_hook = psutil.Process(pid=self._pid)
+        self.enabled = True
 
-    def stop_cow_master(self):
-        self._proc_hook.terminate()
+    def stop_cow_master(self, disable=True):
+        if self._proc_hook:
+            self._proc_hook.terminate()
+            self.enabled = disable
+            self._pid = None
+            self._proc_hook = None
     
     async def set_client_connection(self, client_connection):
         LOGGER.highlight("CowMaster - Connected to manager.")
         self.client_connection = client_connection
+        self._proc_hook = MISC.get_client_pid_by_tcp_source_port(self.global_config['hon_data']['svr_managerPort'], client_connection.addr[1])
+        self._pid = self._proc_hook.pid
+    
+    def unset_client_connection(self):
+        self.client_connection = None
+        self._pid = None
+        self._proc_hook = None
 
     async def on_game_state_change(self, key, value):
         # do things
@@ -65,9 +79,41 @@ class CowMaster:
 
     def get_port(self):
         return self.port
-
-    def reset_game_state(self):
-        LOGGER.debug(f"GameServer #{self.id} - Reset state")
+    
+    def reset_cowmaster_state(self):
+        LOGGER.debug(f"CowMaster - Reset state")
         self.status_received.clear()
         self.game_state.clear()
-        self.game_in_progress = False
+    
+    async def monitor_process(self):
+        LOGGER.debug(f"CowMaster - Process monitor started")
+        try:
+            while not stop_event.is_set():
+                if self._proc_hook is not None:
+                    try:
+                        status = self._proc_hook.status()  # Get the status of the process
+                    except psutil.NoSuchProcess:
+                        status = 'stopped'
+                    if status in ['zombie', 'stopped'] and self.enabled:  # If the process is defunct or stopped. a "suspended" process will also show as stopped on windows.
+                        LOGGER.warn(f"CowMaster stopped unexpectedly")
+                        self._proc_hook = None  # Reset the process hook reference
+                        self._pid = None
+                        self._proc_owner = None
+                        self.reset_cowmaster_state()
+                        # the below intentionally does not use self.schedule_task. The manager ends up creating the task.
+                        await self.start_cow_master()
+                    elif status != 'zombie' and not self.enabled:
+                        #   Schedule a shutdown, otherwise if shutdown is already scheduled, skip over
+                        self.stop_cow_master()
+
+                for _ in range(5):  # Monitor process every 5 seconds
+                    if stop_event.is_set():
+                        break
+                    await asyncio.sleep(1)
+
+        except asyncio.CancelledError:
+            LOGGER.debug(f"GameServer #{self.id} Process monitor cancelled")
+            # Propagate the cancellation
+            raise
+        except Exception as e:
+            LOGGER.error(f"GameServer #{self.id} Unexpected error in monitor_process: {e}")
